+++ src/Troschuetz.Random/Generators/AbstractGenerator.cs
@@ -488,22 +488,6 @@ namespace Troschuetz.Random.Generators
         ///   A double-precision floating point number greater than or equal to 0.0, and less than
         ///   1.0; that is, the range of return values includes 0.0 but not 1.0.
         /// </returns>
-        protected static unsafe double ToDouble(uint value)
-        {
-            const double UIntToDoubleMultiplier = 1.0 / (uint.MaxValue + 1.0);
-            return value * UIntToDoubleMultiplier;
-        }
-
-        /// <summary>
-        ///   Converts given <paramref name="value"/> into a double-precision floating point number
-        ///   greater than or equal to 0.0, and less than 1.0; that is, the range of return values
-        ///   includes 0.0 but not 1.0.
-        /// </summary>
-        /// <param name="value">The value that needs to be converted.</param>
-        /// <returns>
-        ///   A double-precision floating point number greater than or equal to 0.0, and less than
-        ///   1.0; that is, the range of return values includes 0.0 but not 1.0.
-        /// </returns>
         protected static unsafe double ToDouble(ulong value)
         {
             // 12 bits are discarded because double mantissa is 52 bits long. The exponent on the